[mobile] Add Device Memory to the device adaption section
The Device Memory specification makes it easier to classify devices to
perform server-side adaptation.

Notes:
- The spec could perhaps also be mentioend in the Performance section (not done
here), but it definitely needs to appear in device adaptation.
- The spec also defines an API, so technically it's also for "JS-based"
adaptation. The API is mentioned in the text but I stuck to server-side
adaptation for the name of the feature to keep the messaging simple.