[Media] Mention target latency exposed in captured content
Media capture page update to clarify that the Media Capture and Streams also
exposes information about capture devices, along with a few examples including
latency.

Orchestration page also updated to add Media Capture and Streams and insist on
the targeted latency exposed by the spec, which allows to relate timestamps in
the captured content to events happening in the real world.

See related suggestion in:
https://lists.w3.org/Archives/Public/public-web-and-tv/2017Oct/0005.html