[Framework] Expose impl sources and complete impl status rules (#109)
Update the Python code that computes the implementation status of each feature
to also expose the implementation data and sources used to compute the final
status in an "implementations" property. That information is not used for the
time being, but could typically be used later on to complete the information in
the implementation column of generated tables.

The rules used to compute the implementation status have also been refined not
to end up in situations where we claim that a feature has shipped in Safari and
is still in development in Webkit (as reported in #87).

The rules are the following (rules appear as code comments too):
1. Trust platform sources to say the right thing about their own user-agent or
rendering engine. For instance, if chromestatus says that a feature is
"in development" in Chrome, consider that the feature is really "in development"
in Chrome, and ignore possible claims in other sources that the feature is
"shipped" in Chrome
2. Keep the most optimistic status otherwise, meaning that if chromestatus says
that feature A has shipped in Edge while caniuse says it is in development,
consider that the feature has shipped in Edge
3. Due to the close relationship between webkit and Safari, trust webkitstatus
more than any other source about support in Safari. If webkitstatus says that a
feature is in development in webkit, it means it cannot be at a more advanced
level in Safari. In other words, constrain the implementation status in Safari
to the implementation status in Webkit, when it is known to be lower.
4. Also, once 3. is done, drop the Webkit entry when there is also an entry for
Safari. No need to confuse people with the distinction between Safari and
Webkit, unless that is needed (needed for features that are being developed in
Webkit but for which there has not been any public signal about support in
Safari)

(Update also switches the Caniuse URL to the HTTPS version)