Add icons to navigation menu
The side menu now displays icons. The same grey filter as the one applied to
home page icons on narrow screens is applied to icons in the side navigation
menu so that the menu remains "gentle to the eye".

Also, the menu hamburger is now automatically hidden in pages when there are
no other pages in the roadmap.

Fixes #71.