[Framework] Add titles to well-known section types when not already set
To guarantee consistency of section titles across pages and avoid duplicating
efforts for translations, the framework now automatically sets the title of
sections when the title is missing and when the section is a "well-known" one.

Well-known sections for now are those that have one of the following classes:
- well-deployed
- in-progress
- exploratory-work
- not-covered
- discontinued

Translations of these section titles may be provided in a `sections` property
of the `translations.xx.json`, using the above class names as keys and
translated titles as values.

This update addresses issue #69.