[Framework] Improve i18n code (mostly "editorial" updates)
* Moved English labels that were hardcoded in the code to a separate
`js/translations.json` file, to improve symmetry between languages.
* Within translations files, moved implementation status labels to a separate
`implstatus` property, so that `labels` only contains real labels.
* Renamed `groupnames` and `spectitles` properties into `groups` and
`specifications` to avoid having to think about whether to use camel case, and
improve consistency with other property names.
* Improved localization warnings reported to the console, to avoid duplicates
in particular
* Implemented an ad-hoc mechanism to deal with the "in" label, which is now
"%feature in %spec". This should (hopefully) be enough for now.
* Completed the translation instructions in the README, to explain which files
need to be translated and where to put them.

This commit addresses the following issues:
* Update translation instructions in README (#118)
* Add templating mechanism for translation strings (#114) (although note I used
an ad-hoc mechanism using calls to `replace`, but we should not have many such
strings to translate, so that seems good enough for now)