[Framework] Add more flexible table definition mechanism
This update gets rid of the HTML table templates that did not serve any useful
purpose, since the logic that goes with a table needs to be implemented in
the code one way or the other. Said differently, the code needs to know the
structure of the tables it is to generate, so hiding that structure in an HMTL
template is useless and error-prone.

The table column headers now need to appear in translation files in a `columns`
property (see `translations.json` for reference).

The types of tables and the columns to display for each of them can now be
customized on a roadmap per roadmap basis, in the `toc.json` file, in a
`tables` property that lists the columns to use per type of table. This
mechanism can be used to override the default tables or to create new types
of tables.

For instance, to add the "group" column to the columns generated in
"well-deployed" sections, one can simply add the following to the `toc.json`
file:

```json
{
  "tables": {
    "well-deployed": ["feature", "spec", "group", "maturity", "impl"]
  }
}
```

I created a new "versions" type of column, which should eventually become what
@r12a needs in #90 (not quite the case now because the framework does not
yet know how to gather the Editor's Draft and the GitHub repo of a spec)