[Framework] Impl and polyfills can now be listed in data/ files
The implementation status can now be specified in the data/ file that describes
a feature, in an `other` property of an `impl` property. That property should
take an object that sets the additional implementation status per user agent.

For instance:
```json
{
  "impl": {
    "other": {
      "safari": "shipped",
      "chrome": "indevelopment",
      "firefox": "experimental"
    }
  }
}
```

Alternatively (but not necessarily encouraged for now as long as information
remains simple), the property can take an array of implementation info as in:

```json
{
  "impl": {
    "other": [
      { "ua": "safari", "status": "indevelopment", "source": "rumors" },
      { "ua": "edge", "status": "experimental", "source": "dreams" }
    ]
  }
}
```

This implementation info completes the implementation info gathered
automatically in `spec/impl.json`.

Polyfills can now be listed in the data/ file as well, in a `polyfills`
property. The value of that property should be an array of objects with a
`url` and a `label` property (possibly an `img` property as well). That info
gets copied over to `spec/impl.json` and rendered in the implementations
column of generated tables next to user-agent implementations.

The update addresses the following two issues:
- #7 Some way to specify implementation status locally
- #84 Possibility to list JS libraries as implementations