[Data] Drop ref to "Can I Use" for CSS Backgrounds and Borders
The reference was to CSS border radius, which was incorrect. There are no
entries for the CSS Background and Borders spec as such on any of the status
platforms.

There are entries for individual properties of the CSS Backgrounds and Borders
spec in Can I Use, but we still do not have a proper way to reference them
here:
- background-attachment
- css-backgroundblendmode
- css-background-offsets
- background-repeat-round-space

Fixes #132.