Drop duplicated JS logic
This commit removes the need to link to a separate `generate-index.js` script
for menu pages, merges the index and regular page HTML templates, and moves
function declarations to a `utils.js` file to improve readability of the
generation script.

This commit addresses issue #113 by actually removing the need to maintain
different generation logic for menu pages and for regular pages.