Auto-link between translations of a roadmap page
Provided the `toc.json` file contains a `translations` key that lists existing
translations, the framework will now automatically generate the links between
translations in the footer of the page.

The `translations` key should follow the following structure:

```json
{
  "translations": [
    {
      "title": "English",
      "lang": "en"
    },
    {
      "title": "中文",
      "lang": "zh"
    }
  ]
}
```

The list of translations does not need to appear in localized versions of the
TOC file, because the framework now loads the default `toc.json` file no matter
what and merely overwrites the properties it defines with those it finds in
the localized version of the file `toc.xx.json`. This also has the benefits that
content no longer needs to be duplicated between `toc.json` file and
`toc.xx.json`.

Note the framework now automatically generates the "right" localized links in
the navigation menu, e.g. given an English roadmap page named "storage.html",
the framework knows that the Chinese version will be "storage.zh.html", and
will link to that page in the navigation menu.