Filter UA in implementation status column based on query string param (#165)
The new `filter-implstatus.js` script runs once a page has been generated
(either dynamically or statically if the current page is a published export),
and filters the list of user agents that appear in "Current implementations"
columns in generated tables based on the presence of a "ua" query string
parameter.

The script expects the query string parameter to contain a comma-separated list
of the user-agents that need to appear. For instance:
- `ua=firefox` to only display the current implementation status in Firefox
- `ua=chrome,edge` to display the current implementation status in Chrome and
in Edge.

The script also updates navigation links to preserve the query string across
pages.