[Data] Force Webkit status for Presentation API (#182)
See #173. The Chrome status platform incorrectly reports that the Presentation
API is in development in Safari. This is a priori wrong. This update forces the
Webkit implementation status to `""`, meaning to "not implemented".

Note the Python code that computed the Webkit status ignored the empty string
when it bumped into one. It now makes a difference between `""` (which means
"not implemented"), and `None` (which means "no info").