[Mobile] Mark humidity, thermometer, barometer proposals as discontinued (#187)
See #159. The roadmap mentioned the Ambient Humidity Events spec in the
Exploratory Work section but work has been discontinued. This update moves
that spec to the Discontinued feature section, and completes the note to
mention the Thermometer and Barometer proposals. Work on these proposals may
resume any time if there is enough implementer support.