Fix Ambient light API implementation status in Firefox and Edge (#186)
* Fix Ambient light API implementation status in Firefox and Edge

See #160. I completed the implementation algorithm to consider that a "feedback"
source is authoritative. I also added a comment and a date to the feedback
information in the data folder to ease maintenance over time. The Python code
passes that information over to the JavaScript, which does not do anything with
it for the time being.

* Implementation status description in README file

The update moves the text that describes the `impl` property to a dedicated
sub-section, which now also explains in more detail how to enter implementation
status manually.