[Mobile] Update accessibility part in User input
See #184.

Some of the references were (still mostly valid but) outdated. The text was
updated and is much more complete, mentioning the different guidelines
(UAAG for user agents, WCAG for developers, ATAG for authoring tools), and
highlighting mobile-oriented specificities when they exist.

The "User Input" section was also renamed into "User Interaction", which seems
a better term to describe what the page is about. The page description mentions
accessibility as well, and the page now starts with an introduction that
explains the scope of the page.