[Framework] Convert Python scripts to Node.js (#206)
This update converts the extract-spec-data and extract-impl-data scripts from
Python to Node.js. Node.js v8.0.0 or above is required. "npm install" has to
be run before the scripts can be run.

Goal is to get rid of Python altogether to ease maintenance and help with
future improvements that we'll want to make to these scripts (no offence meant
to Python, I'm simply more familiar with Node.js these days).

Scripts produce the same files as before, except the order of properties for
object can be slightly different in the resulting JSON. The extract-impl-data
script in Python actually had an error that made him report incorrect Edge
implementation info (in particular for CSS Media Queries 4).

No attempt has been made to improve the logic at this stage. It's just meant to
be a 1-to-1 replacement of the scripts in Python.

Makefile updated and deploy script for Travis CI updated as well.

NB: The Node.js script make use of a little library to fetch URLs and save
the results into a cache folder (".cache" by default). This helps speed up
future "makes", and allows to bypass network requests entirely for testing
purpose if you create a "config.json" file that contains an
"avoidNetworkRequests" property set to true. See corresponding doc for details:
https://github.com/tidoust/fetch-filecache-for-crawling#configuration