Leave W3C API key in plain text to allow build checks in PR
Travis CI removes encrypted environment variables on pull requests from a fork
for security reasons:
https://docs.travis-ci.com/user/pull-requests/#Pull-Requests-and-Security-Restrictions

This would effectively prevent checking pull requests before they get merged.
Or we'd have to send PR from the same repository, which I'm not a huge fan of.

The W3C API allows leaving the W3C API key in plain text, so let's stick to that
for now. I'll try to restrict the key to the right set of domain names later on
(not exactly sure how to do that for now)