Send Origin HTTP header to W3C API server
Since the W3C API key appears in plain text in `.travis.yml` (to allow checks
to run in pull requests from forks of the repo), it is highly recommended to
restrict the API key to a specific set of domain names.

The code now sends an `Origin: https://www.w3.org` HTTP header, and the key was
restricted to `w3.org`. This should prevent reuse of the API key in contexts
that automatically set the `Origin` header (such as in browser contexts).