[Framework] Fetch spec info from W3C API and Specref (#207)
Major re-write of the spec info extraction script to fetch info from the W3C API
and from Specref. A valid W3C API key is required to query the W3C API. It must
be written to a `w3cApiKey` property in a `config.json` file in the root folder
or as a `W3C_API_KEY` environment variable.

Also took the opportunity to inject all (non implementation related) info
defined in data files into the generated `tr.json` file so that loading
individual data files from the Web page becomes useless. That simplifies the
logic of the JS generation a bit.

The code also drops the need to distinguish between the `TR`, `editors` and `ls`
URLs of a spec in the data file. A `url` property should be enough if the spec
exists in the W3C spec database or in Specref (the code will also consider that
the filename is a W3C shortname if that property does not exist to generate the
spec's URL). README updated accordingly. The code still supports the former
properties.

Data files have not yet been updated to use `url` instead of the former
properties. Most of the `wgs` and `title` properties can also be dropped now
that the code fetches info for Specref.

The Editor's Draft URL is now also included in `tr.json`. The UI does not make
use of that info for now.

Note that Travis CI removes encrypted environment variables on pull requests
from a fork for security reasons:
https://docs.travis-ci.com/user/pull-requests/#Pull-Requests-and-Security-Restrictions

As a consequence, the W3C API key used by Travis CI appears in plaintext in
`.travis.yml`. That key is restricted to requests that contain an `Origin` HTTP
header with a `w3.org` domain name.