[Framework] Improve group info logic and expose publisher
Addresses #208.

The code now makes use of the `publisher` property returned by Specref to
find a friendly label for the group that delivered the spec.

It also tries to set a `publisher` property, which is for now one of W3C, IETF
or WHATWG. If that property is not already set, it looks at the URL of the
group and at the URL of the spec to find out which organization published it.