[Framework] Complete impl info (prefix, flag, notes)
This update completes the implementation info retrieved from Web platform status
sources:
- a `prefix` flag is set when the implementation requires a prefix
- a `flag` flag is set when the implementation is behind some sort of flag
- status is set to `experimental` instead of `shipped` whenever the `prefix` or
`flag` flags are set
- a `notes` array contains implementation notes when available (the `comment`
property in the data file gets mapped to that array)
- mobile browser names normalized in a more logical way, now follow the names
used in mdn-browser-compat (preliminary work for #13)
- better support for Chrome implementation info, in particular on Android
(preliminary work for #13)

This info is not yet being used by the Web application, so the update does not
change much for now, except for:
- a couple of specs that were reported as shipped and that are now reported as
"experimental", e.g. Web Crypto in Edge which still requires a prefix
- a couple of specs that were reported as shipped in Chrome and that are now
no longer reported because implementation is actually in Chrome for Android
(e.g. the Remote Playback API)

Note that MSE support in Chrome now seems wrong to me (experimental instead of
shipped), but the thing is Chrome status reports that a prefix is still
required:
https://www.chromestatus.com/features/4563797888991232