[Framework] Fix group name rendering
Group name is now returned in `tr.json` in a `deliveredBy` property, but the
code in `utils.js` had not been updated. This worked in some cases because
`tr.json` still contained a `wgs` property when that property was set in the
data file, but group name was not rendered for specs retrieved from the W3C API
or from Specref.