[Framework] Complete implementation info with badges
The code now renders more specific implementation information on screen using
visual badges on top of browser icons. Badges are:
- Feature available on desktop only / mobile only / desktop and mobile
- Features requires using a prefix
- Feature is behind a flag

Essentially, there will always be a desktop/mobile badge and, most of the time,
that badge will be the "desktop and mobile" one, because most recent features
are developed on all platforms.

A tooltip explains what the implementation info means, for instance:
"Experimental in Chrome (mobile). Feature requires using a prefix. Source:
Chrome Platform Status."

That text also appears in a span element for accessibility readers.

Clicking on the browser icon takes the reader to the right page on the Platform
Status Web site where the implementation info comes from.

Some changes made to the `extract-impl-data.js` script to make things possible:
- The script now sets a `selected` flag on the implementation info that is
considered to be the right one for each user agent.
- The script also sets the URL to look at on the underlyin Platform Status Web
site to get additional info.
- The script only runs the selection logic once (it incorrectly ran after each
step)
- The Chrome Platform Status platform was marked as authoritative for Chrome
for Android (it was only considered as authoritative for Chrome for Desktop)

The styles of the generated table have been slightly updated to set some
minimum width to the implementation column (enough to render 5 browsers),
and some maximum width to the feature column (to keep it short).

Also created a `tr` function to translate strings instead of passing the
translations object around.