[Framework] Write spec title and group name in same column
To save some width in generated table for implementation info, and keep all of
our tables consistent, the name of the group(s) responsible for a given spec
or feature is (are) listed below the specification title in the same column.

Styles used are similar to those in the new /TR/ page, the group name appearing
in light grey below using a smaller font size.

Links now use consistent styles throughout the roadmap, regardless of whether
they appear in the body of the text or in generated tables. No underline in
the tables in particular to improve readability.

The update also removes and sanitizes a few styles that were defined twice in
`theme.css` and `roadmap.css`. More work would be needed on that front, but
we'll do that later on (for once, I'm not sure why we have two files).

Updates to link styles meant the remaining `<strong><a>...</a></strong>`
constructs rendered in bold again. I dropped the `<strong>` element everwhere.