[Framework] Translation logic for implementation filtering
The implementation filtering mechanism now uses the JSON translation files to
display the filtering menu label and the browser names.

Code re-shuffled accordingly to put JavaScript functions common to the
generation script and to the filtering script in the `utils.js` script and
put JavaScript functions that are only useful for the generation script in a
separate `generate-utils.js` script.

Translation function also updated to fallback to English translations in all
cases and record all missing translations in a global `warnings` array
displayed when the page is generated at the end of the scripts' initialization
to help with translations.