[Framework] Move filtering menus generation logic
The filtering menus of the implementation column are now generated by the
`generate.js` script. This means that the `filter-implstatus.js` script:
1. stays agnostic about browsers that it needs to filter. The script merely
uses the info in the filtering menus and does not need to hardcode a precise
list of browsers.
2. no longer needs to load translation files, which thus do not need to be kept
around in published snapshots.

The `generate-utils.js` and `utils.js` scripts have been kept separated for now
to ease review but they can be merged again afterwards.

The update also adds the possibility to provide translations of "desktop" and
"mobile" in the tooltip.