[Framework] Add an about page to all roadmaps (#221)
See #122. The update adds an about page to all existing multi-page roadmaps
using a default page for now, common to all.

That default page describes the goals and audience of the document, and contains
a legend for generated tables that explains the meaning of the different icons
and badges.

The framework embeds a mechanism that allows authors to override content in
the default about template. The README file was updated accordingly.

The `about.zh.html` file will obviously need some additional love, and a
`template-about.zh.html` file will have to be created.

Some styles adjusted for tables. Also, note maturity icons now all have the
same size, so the update drops the offending code that set the height
differently depending on the icon (#212).