[Framework] A series of UI accessibility improvements
This update contains a short series to address main reported accessibility
issues of the User Interface. Changes are:

- Links in text are underlined (#226)
- Menu is now keyboar accessible and "Tab/Up/Down/Home/End" key friendly (#227)
- Feature labels in tables are top-aligned when there the column spans multiple
rows (#229)
- Menu button is correctly labelled (#230)
- Contrast of table headers improved (#231)
- Contrast of footer improved (#232)

Actually, the menu can now be opened and closed through the "m" key. I haven't
tried to advertise that anywhere for now.

The update also fixes the URL of the about page in translations (#233)