[Framework] Script to generate all roadmap pages automatically (#235)
This update adds a script that generates all roadmap pages to a `.out` folder
and copies relevant assets there as well. Addresses #129 (and part of #228 in
the sense that we'll be able to replace the `gh-pages` branch with the
generated pages).

Under the hoods, the script uses JSDOM to load HTML pages, run the generation
in individual pages, and save the generated HTML content.

Some bits of code were adjusted accordingly not to create absolute URLs for
translation links, and to respond to the right "load" event in generated
documents.