[Framework] Deploy fully generated pages
Fix for #129.

The pages published on GitHub so far were the semi-generated pages: the feature
info and implementation info were generated after each push but the pages were
assembled on the client device.

With this update, pages in the `gh-pages` branch become the fully generated
pages, meaning pages that have already been assembled and that are mostly
static, except for a couple of scripts (one to handle the menu and one to
implement the browser filtering mechanism).

The update fixes the flickering issue (#228) in published pages (technically,
that problem still exists in page authoring mode, but that seems acceptable).