[Framework] Auto-generate Document Metadata section (#241)
See #166. The framework will now automatically generate a Document Metadata
section in the index page of a multi-page roadmap, with appropriate links.
README updated to mention the different settings that may need to be added to
the `toc.json` file, namely `publishedVersion`, 'edDraft, `thisVersion`,
`previousVersion`, `github` and `publishDate`.

When specified, the `publishDate` appears next to the page's title in all pages.
Other information only appear in the index page.