[Mobile] Add a changelog since previous snapshot (#243)
Add a changes since previous published version to the About page that highlights
the list of updates in different sections, and quickly explains other changes
made to the UI and underlying framework.

This update also introduces a generic way to link to roadmap pages with a
`data-page` attribute whose value must be the URL of the page as it appears
in the `toc.json` file (without the extension). The framework updates these
links automatically, targeting the localized version when needed.