[Framework] Fix generation of Editor's Draft URL from github repo
See #244. The code assumes that the repo will be cloned locally to a folder
whose name is the name of the repo. This is at least true for deployment on
Travis, and should be true in most cases. There aren't good ways to compute
the end of the path otherwise...