[Framework] Add last modified dates to footer (#249)
See #72. Two dates are being mentioned in the footer:
1. The date when the content iself was last updated. The assumption is that the
last modified date of the document will be the correct date.
2. The date when the implementation data (and spec info) was last updated. The
assumption is that this is the current date at the time when the page is
generated.

As an exception to the rule, if `publishDate` is specified, then it is used
(except for the content if it is more recent than `document.lastModified`)

The dates are not perfect. Also, having two dates is probably a bit confusing.
But then both dates seem to make sense, and could be useful, especially on
roadmap pages published on GitHub (e.g. to point out that content may be 6
months old, but implementation data is still "current").