[Framework] Localize date strings in pages
Last modified dates and publication dates were always formatted in English.
They should now be formatted according to the language of the page.

Under the hoods, the `toLocaleDateString` method gets used to format the date.
Note browsers may only support a subset of locale settings in practice (and
may for instance only support locale settings that are actually local to the
user), so that mechanism may not always work.

In particular, JSDOM, used to generate all roadmap pages in batch, does not
seem to support Chinese, so rules for Chinese are hardcoded if JSDOM is
detected.

This commit is a fix for:
https://github.com/w3c/web-roadmaps/issues/213#issuecomment-386197580