[Framework] Custom maturity icons for evergreen/informative specs (#265)
Features initially proposed in https://github.com/w3c/web-roadmaps/issues/90.

W3C specs are automatically flagged as informative when they contain only
informative content or when they will be (or have been) published as a Note.

Specs can also be flagged as "evergreen", which means that, regardless of their
exact status on the Rec-track (or elsewhere), they can be considered fit for
reference. This is roughly equivalent to the notion of a Living Standard in
WHATWG.

These settings are used to adjust the maturity icon being rendered. Notably:
- regardless of their exact maturity status, evergreen specs use a green "REF"
(for "Reference") icon.
- informative icons are in dark grey to convey the fact that the spec does not
define normative content.

Note these colors do not match those proposed in the initial issue. This can be
adjusted later on if needed.