[Framework] Support linking to more than one ID on status platform
See https://github.com/w3c/web-roadmaps/issues/59

The description of a specification can now list features that compose the spec
along with info about where to extract implementation status for each feature
on status platforms.

For instance, to say that the CSS Backgrounds spec contains the features
"Rounded corners" and "Backgrounds", one can define the spec as:

```json
{
  "url": "https://www.w3.org/TR/css-backgrounds-3/",
  "impl": { ... },
  "features": {
    "border-radius": {
      "url": "#the-border-radius",
      "title": "Rounded corners",
      "impl": { ... }
    },
    "background": {
      "url": "#the-background",
      "title": "Backgrounds",
      "impl": { ... }
    }
  },
  "featuresCoverage": "partial"
}
```

The `featuresCoverage` flag is set to `partial` by default. If set to `full`,
and if we have implementation info from one source on all features for a given
browser, then the framework will guess that the browser supports the entire
spec.

This new mechanism replaces the previous `feature` property (that previous
mechanism can still be used, although it is deprecated).

When support may be partial, an orange banner gets displayed on top of the
implementation icon to convey the possibility that the implementation status
may not apply to the whole spec.